requirements.txt
autopep8==1.3.2
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.4
django-admin-env-notice==0.1.4
django-debug-toolbar==1.8
django-decouple==2.1
django-import-export==0.5.1
django-jet==1.0.6
django-otp==0.4.0.1
django-rest-swagger==2.1.2
django-silk==1.0.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
oauth2client==4.1.2
odfpy==1.3.5
openapi-codec==1.3.2
openpyxl==2.4.8
pandas==0.20.3
pkg-resources==0.0.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycodestyle==2.3.1
Pygments==2.2.0
PyJWT==1.5.2
pymaging==0.1
pymaging-png==0.1
python-dateutil==2.6.1
python-logstash==0.4.6
pytz==2017.2
PyYAML==3.12
qrcode==5.3
requests==2.18.4
rsa==3.4.2
simplejson==3.11.1
six==1.10.0
sqlparse==0.2.3
tablib==0.12.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
